Небольшие правки

+++ src/main/java/pro/sky/petshelterbot/service/PetShelterService.java
@@ -4,6 +4,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
 import pro.sky.petshelterbot.entity.Pet;
 import pro.sky.petshelterbot.entity.Shelter;
 import pro.sky.petshelterbot.entity.Volunteer;
@@ -12,18 +14,21 @@ import pro.sky.petshelterbot.repository.PetRepository;
 import pro.sky.petshelterbot.repository.ShelterRepository;
 import pro.sky.petshelterbot.repository.VolunteerRepository;
 
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
-public class DogShelterService {
+public class PetShelterService {
     final private PetRepository petRepository;
     final private ShelterRepository shelterRepository;
     final private VolunteerRepository volunteerRepository;
 
     final private Shelter catShelter, dogShelter;
 
-    public DogShelterService(
+    public PetShelterService(
             PetRepository catRepository,
             ShelterRepository shelterRepository,
             VolunteerRepository volunteerRepository) {
@@ -35,18 +40,46 @@ public class DogShelterService {
         this.volunteerRepository = volunteerRepository;
     }
 
-    public Pet createDog(Pet pet) {
-        if (!pet.getSpecies().equals("dog")) {
-            throw new ShelterException("Проверьте тип животного.");
+    private static String UPLOAD_DIRECTORY = "src/main/resources/uploads/pet_images";
+
+    private String imgUploader(Long id, MultipartFile img) {
+        try {
+            String fileName = String.format(
+                    "%d.%s",
+                    id,
+                    StringUtils.getFilenameExtension(img.getOriginalFilename())
+            );
+            byte[] data = img.getBytes();
+            Path path = Paths.get(UPLOAD_DIRECTORY, fileName);
+            Files.write(path, data);
+            return path.toString();
+        } catch (Exception e) {
         }
-        if (!pet.getShelter().getType().equals("dog")) {
-            throw new ShelterException("Проверьте тип приюта.");
-        }
-        return petRepository.save(pet);
+        return null;
     }
 
-    public Pet add(Pet pet) {
-        return petRepository.save(pet);
+    public Pet add(Pet pet, MultipartFile img) {
+        if (pet.getShelter().getType().equals(pet.getSpecies())) {
+            String imgPath = imgUploader(pet.getId(), img);
+            if (imgPath != null) {
+                pet.setImgPath(imgPath);
+            }
+            return petRepository.save(pet);
+        }
+        throw new ShelterException("Тип животного не соответствует типу размещаемых животных в приюте.");
+    }
+
+    public Pet addImg(Long petId, MultipartFile img) {
+        Pet pet = petRepository.getPetById(petId);
+        if (pet == null) {
+            throw new ShelterException("Питомец не найден.");
+        }
+        String imgPath = imgUploader(petId, img);
+        if (imgPath != null) {
+            pet.setImgPath(imgPath);
+        }
+        petRepository.save(pet);
+        return pet;
     }
 
     public Pet add(String species, String name, Shelter shelter) {
@@ -64,8 +97,10 @@ public class DogShelterService {
         return pet;
     }
 
-    public Pet delete(Long id) {
-        Pet pet = get(id);
+    public Pet delete(Pet pet) {
+        if (petRepository.findById(pet.getId()).isEmpty()) {
+            throw new ShelterException("Животное не найдено.");
+        }
         petRepository.delete(pet);
         return pet;
     }
